requirements.txt
alembic==1.17.0
annotated-types==0.7.0
anyio==4.11.0
async-timeout==5.0.1
asyncpg==0.30.0
bcrypt==4.0.1
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cryptography==46.0.2
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
exceptiongroup==1.3.0
fastapi==0.119.0
greenlet==3.2.4
h11==0.16.0
httptools==0.7.1
idna==3.11
lxml==6.0.2
Mako==1.3.10
MarkupSafe==3.0.3
passlib==1.7.4
pillow==11.3.0
psycopg==3.2.10
psycopg-binary==3.2.10
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.0
pydantic_core==2.41.1
PyMuPDF==1.26.5
python-docx==1.2.0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
PyYAML==6.0.3
reportlab==4.4.4
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.48.0
tomli==2.3.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.37.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
